Share ContainerFlags
