+++ CSSE/CSContainer.h
@@ -2,23 +2,9 @@
 
 #include "CSActor.h"
 
+#include "TES3ContainerFlags.h"
+
 namespace se::cs {
-	namespace ActorFlagContainer {
-		typedef unsigned int value_type;
-
-		enum Flag : value_type {
-			Organic = 0x1,
-			Respawns = 0x2,
-			IsBase = 0x8,
-		};
-
-		enum FlagBig {
-			OrganicBit = 0,
-			RespawnsBit = 1,
-			IsBaseBit = 2,
-		};
-	}
-
 	struct Container : Actor {
 		int unknown_0x7C;
 		int unknown_0x80;
