+++ MWSE/TES3Container.h
@@ -1,26 +1,11 @@
 #pragma once
 
+#include "TES3ContainerFlags.h"
 #include "TES3Defines.h"
 
 #include "TES3Actor.h"
 
 namespace TES3 {
-	namespace ActorFlagContainer {
-		typedef unsigned int value_type;
-
-		enum Flag : value_type {
-			Organic = 0x1,
-			Respawns = 0x2,
-			IsBase = 0x8,
-		};
-
-		enum FlagBig {
-			OrganicBit = 0,
-			RespawnsBit = 1,
-			IsBaseBit = 2,
-		};
-	}
-
 	struct ContainerBase : Actor {
 		static constexpr auto OBJECT_TYPE = ObjectType::Container;
