+++ SharedSE/SharedSE.vcxitems.filters
@@ -375,6 +375,9 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3CreatureFlags.h">
       <Filter>Header Files\TES3</Filter>
     </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3ContainerFlags.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)NIMatrix33.cpp">
